+++ mongod/structs.go
@@ -31,10 +31,10 @@ type AppRepository interface {
 	CreateArch(archName string, owner string, ctx context.Context) (interface{}, error)
 	ListArchs(ctx context.Context, owner string) ([]*model.Arch, error)
 	DeleteArch(id primitive.ObjectID, owner string, ctx context.Context) (int64, error)
-	CreateApp(appName string, logo string, description string, private bool, tuf bool, reports bool, owner string, ctx context.Context) (interface{}, error)
+	CreateApp(appName string, logo string, description string, private bool, tuf bool, reports bool, cdnEdge bool, owner string, ctx context.Context) (interface{}, error)
 	ListApps(ctx context.Context, owner string) ([]*model.App, error)
 	DeleteApp(id primitive.ObjectID, owner string, ctx context.Context) (int64, error)
-	UpdateApp(id primitive.ObjectID, appName string, logo string, tuf bool, description string, reports bool, owner string, ctx context.Context) (interface{}, error)
+	UpdateApp(id primitive.ObjectID, appName string, logo string, tuf bool, description string, reports bool, cdnEdge bool, owner string, ctx context.Context) (interface{}, error)
 	UpdateChannel(id primitive.ObjectID, paramValue string, owner string, ctx context.Context) (interface{}, error)
 	UpdatePlatform(id primitive.ObjectID, platformName string, updaters []model.Updater, owner string, ctx context.Context) (interface{}, error)
 	UpdateArch(id primitive.ObjectID, paramValue string, owner string, ctx context.Context) (interface{}, error)
@@ -75,6 +75,7 @@ type Changelog struct {
 type CheckResult struct {
 	Found                  bool
 	Critical               bool
+	CdnEdge                bool
 	Artifacts              []Artifact
 	Changelog              []Changelog
 	IsRequiredIntermediate bool