+++ server/utils/storage/interfaces.go
@@ -11,6 +11,7 @@ type StorageClient interface {
 	UploadObject(ctx context.Context, bucketName, objectKey string, fileReader multipart.File, contentType string) error
 	UploadPublicObject(ctx context.Context, bucketName, objectKey string, fileReader multipart.File, contentType string) (string, error)
 	DeleteObject(ctx context.Context, bucketName, objectKey string) error
+	DeleteObjects(ctx context.Context, bucketName string, objectKeys []string) error
 	GeneratePresignedURL(ctx context.Context, bucketName, objectKey string, expiration time.Duration) (string, error)
 	DownloadObject(ctx context.Context, bucketName, objectKey string, filePath string) error
 	ListObjects(ctx context.Context, bucketName, prefix string) ([]string, error)